requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
django-cors-headers==3.13.0
Pillow==9.3.0